+++ tests/test_cli.py
@@ -74,7 +74,7 @@ def test_cli_create():
 def test_cli_create_unsupported(svg_file):
     result = runner.invoke(app, ["create", svg_file])
     assert result.exit_code == 1
-    assert "Unsupported mediatype image/svg+xml" in result.stdout
+    assert "No known processing mode for image/svg+xml" in result.stdout
 
 
 def test_cli_batch_no_arg():